test: optimize st test
