+++ tests/basic/tools/test_shell_tool.py
